Cleaned up the freeverb code and removed duplicate code using inheritance
